alter static adapter to be per room.. oops
