+++ room-app/src/main/java/net/wasdev/gameon/room/RoomWS.java
@@ -17,17 +17,13 @@ package net.wasdev.gameon.room;
 
 import java.io.IOException;
 import java.io.StringReader;
-import java.util.Collection;
-import java.util.HashSet;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
-import java.util.Map.Entry;
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.ConcurrentHashMap;
 
 import javax.json.Json;
-import javax.json.JsonArrayBuilder;
 import javax.json.JsonObject;
-import javax.json.JsonObjectBuilder;
 import javax.websocket.CloseReason;
 import javax.websocket.Endpoint;
 import javax.websocket.EndpointConfig;
@@ -45,162 +41,86 @@ import net.wasdev.gameon.room.engine.Room;
  */
 public class RoomWS extends Endpoint{
 	private final Room room;
+	private final LifecycleManager.SessionRoomResponseProcessor srrp;
 	
-	public RoomWS(Room room){
+	public RoomWS(Room room, LifecycleManager.SessionRoomResponseProcessor srrp){
 		this.room=room;
-        room.setRoomResponseProcessor(srrp);
+		this.srrp=srrp;
 	}
 	
-	public static class SessionRoomResponseProcessor implements net.wasdev.gameon.room.engine.Room.RoomResponseProcessor{
-		AtomicInteger counter = new AtomicInteger(0);
-		
-	    private void generateEvent(Session session, JsonObject content, String userID, boolean selfOnly, int bookmark) throws IOException {
-	    	JsonObjectBuilder response = Json.createObjectBuilder();
-	    	response.add("type", "event");
-	    	response.add("content", content);
-	    	response.add("bookmark", bookmark);
-	    	
-	    	session.getBasicRemote().sendText("player," + (selfOnly?userID:"*") + "," + response.build().toString());
-	    }
-		public void playerEvent(String senderId, String selfMessage, String othersMessage){
-			//System.out.println("Player message :: from("+senderId+") onlyForSelf("+String.valueOf(selfMessage)+") others("+String.valueOf(othersMessage)+")");
-			JsonObjectBuilder content = Json.createObjectBuilder();
-			boolean selfOnly=true;
-			if(othersMessage!=null && othersMessage.length()>0){
-				content.add("*", othersMessage);
-				selfOnly=false;
-			}
-			if(selfMessage!=null && selfMessage.length()>0){
-				content.add(senderId, selfMessage);
-			}
-			JsonObject json = content.build();
-			int count = counter.incrementAndGet();
-			for(Session s : activeSessions){
-				try{
-					generateEvent(s, json, senderId, selfOnly, count);
-				}catch(IOException io){
-					throw new RuntimeException(io);
-				}
-			}		
+	private Map<Session,MessageHandler> handlersBySession = new ConcurrentHashMap<Session,MessageHandler>();
+	
+	private static class SessionMessageHandler implements MessageHandler.Whole<String> {
+		private final Session session;
+		private final RoomWS owner;
+		public SessionMessageHandler(Session session, RoomWS owner){
+			this.session=session;
+			this.owner=owner;
 		}
-	    private void generateRoomEvent(Session session, JsonObject content, int bookmark) throws IOException {
-	    	JsonObjectBuilder response = Json.createObjectBuilder();
-	    	response.add("type", "event");
-	    	response.add("content", content);
-	    	response.add("bookmark", bookmark);
-	    	session.getBasicRemote().sendText("player,*," + response.build().toString());
-	    }
-		public void roomEvent(String s){
-			//System.out.println("Message sent to everyone :: "+s);
-			JsonObjectBuilder content = Json.createObjectBuilder();
-			content.add("*", s);
-			JsonObject json = content.build();
-			int count = counter.incrementAndGet();
-			for(Session session : activeSessions){
-				try{
-					generateRoomEvent(session, json, count);
-				}catch(IOException io){
-					throw new RuntimeException(io);
-				}
-			}	
-		}
-		public void chatEvent(String username, String msg){
-			JsonObjectBuilder content = Json.createObjectBuilder();
-			content.add("type", "chat");
-			content.add("username", username);
-			content.add("content", msg);
-			content.add("bookmark", counter.incrementAndGet());
-			JsonObject json = content.build();
-			for(Session session : activeSessions){
-				try{
-					session.getBasicRemote().sendText("player,*,"+json.toString());
-				}catch(IOException io){
-					throw new RuntimeException(io);
-				}
+		@Override
+		public void onMessage(String message){
+			try{
+				owner.receiveMessage(message, session);
+			}catch(IOException io){
+				System.err.println("IO Exception sending message to session "+io);
 			}
-		}
-		public void locationEvent(String senderId, String roomName, String roomDescription, Map<String,String> exits, List<String>objects, List<String>inventory){
-			JsonObjectBuilder content = Json.createObjectBuilder();
-			content.add("type", "location");
-			content.add("name", roomName);
-			content.add("description", roomDescription);
-			
-			JsonObjectBuilder exitJson = Json.createObjectBuilder();
-			for( Entry<String, String> e : exits.entrySet()){
-				exitJson.add(e.getKey(),e.getValue());
-			}	
-			content.add("exits", exitJson.build());
-			JsonArrayBuilder inv = Json.createArrayBuilder();
-			for(String i : inventory){
-				inv.add(i);
-			}
-			content.add("pockets", inv.build());
-			JsonArrayBuilder objs = Json.createArrayBuilder();
-			for(String o : objects){
-				objs.add(o);
-			}
-			content.add("objects", objs.build());
-			content.add("bookmark", counter.incrementAndGet());
-			
-			JsonObject json = content.build();
-			for(Session session : activeSessions){
-				try{
-					session.getBasicRemote().sendText("player,"+senderId+","+json.toString());
-				}catch(IOException io){
-					throw new RuntimeException(io);
-				}
-			}
-		}
-		public void exitEvent(String senderId, String message, String exitID){
-			JsonObjectBuilder content = Json.createObjectBuilder();
-			content.add("type", "exit");
-			content.add("exitId", exitID);
-			content.add("content", message);
-			content.add("bookmark", counter.incrementAndGet());
-			JsonObject json = content.build();
-			for(Session session : activeSessions){
-				try{
-					session.getBasicRemote().sendText("playerLocation,"+senderId+","+json.toString());
-				}catch(IOException io){
-					throw new RuntimeException(io);
-				}
-			}
-		}
-		
-		Collection<Session> activeSessions = new HashSet<Session>();
-		public void addSession(Session s){
-			activeSessions.add(s);
-		}
-		public void removeSession(Session s){
-			activeSessions.remove(s);
 		}
 	}
-	
-	private static SessionRoomResponseProcessor srrp = new SessionRoomResponseProcessor();	
-
     @OnOpen
     public void onOpen(final Session session, EndpointConfig ec) {
         // (lifecycle) Called when the connection is opened
-        Log.endPoint(this, "This room is starting up");
         srrp.addSession(session);
+    
+		MessageHandler handlerForSession = new SessionMessageHandler(session, this);
+		
+		MessageHandler fromMap = handlersBySession.get(session);
+        MessageHandler chosen = fromMap!=null?fromMap:handlerForSession;
+        handlersBySession.put(session,chosen);
+        session.addMessageHandler(chosen);
         
-        session.addMessageHandler(new MessageHandler.Whole<String>() {
-			@Override
-			public void onMessage(String message){
-				try{
-					receiveMessage(message, session);
-				}catch(IOException io){
-					System.err.println("IO Exception sending message to session "+io);
-				}
-			}
-		});
+        Log.endPoint(this, "onOpen called against room "+this.room.getRoomId());
+        for(Session s : srrp.activeSessions){
+        	Log.endPoint(this, " Session: "+s.getId());
+        	Log.endPoint(this, "   user? - "+s.getUserProperties().get(Constants.USERNAME));
+        	Log.endPoint(this, "   handlers: "+s.getMessageHandlers().size());
+        	int mhc=0;
+        	for(MessageHandler m : s.getMessageHandlers()){
+        		if(m instanceof SessionMessageHandler){
+        			SessionMessageHandler smh = (SessionMessageHandler)m;
+        			Log.endPoint(this, "    ["+mhc+"] SessionMessageHandler for session "+smh.session.getId()+" linked to room "+smh.owner.room.getRoomId());
+        		}else{
+        			Log.endPoint(this, "    ["+mhc+"] unknown handler");
+        		}
+        		mhc++;
+        	}
+        }
+
     }
 
     @OnClose
     public void onClose(Session session, CloseReason reason) {
         // (lifecycle) Called when the connection is closed, treat this as the player has left the room
-        Log.endPoint(this, "A player has left the room");
         srrp.removeSession(session);
+        MessageHandler handler = handlersBySession.remove(session);
+        if(handler!=null){
+        	session.removeMessageHandler(handler);
+        }
+        
+        Log.endPoint(this, "onClose called against room "+this.room.getRoomId());
+        for(Session s : srrp.activeSessions){
+        	Log.endPoint(this, " Session: "+s.getId());
+        	Log.endPoint(this, "   handlers: "+s.getMessageHandlers().size());
+        	int mhc=0;
+        	for(MessageHandler m : s.getMessageHandlers()){
+        		if(m instanceof SessionMessageHandler){
+        			SessionMessageHandler smh = (SessionMessageHandler)m;
+        			Log.endPoint(this, "    ["+mhc+"] SessionMessageHandler for session "+smh.session.getId()+" linked to room "+smh.owner.room.getRoomId());
+        		}else{
+        			Log.endPoint(this, "    ["+mhc+"] unknown handler");
+        		}
+        		mhc++;
+        	}
+        }
     }
 
     @OnMessage
@@ -218,6 +138,7 @@ public class RoomWS extends Endpoint{
     		removePlayer(session, contents[2]);
     		return;
     	}
+    	System.out.println("ERR: Unknown message type for room "+room.getRoomId()+" message:"+message);
     }
     
     //process a command 
@@ -225,7 +146,7 @@ public class RoomWS extends Endpoint{
 		Log.endPoint(this, "Command received from the user, " + this);
 		JsonObject msg = Json.createReader(new StringReader(json)).readObject();
 		
-		String content = Message.getValue(msg.get("content").toString().toLowerCase());
+		String content = Message.getValue(msg.get("content"));
 		String userid = Message.getValue(msg.get(Constants.USERID));
 		String username = Message.getValue(msg.get(Constants.USERNAME));
 		
@@ -239,20 +160,20 @@ public class RoomWS extends Endpoint{
     
     //add a new player to the room
     private void addNewPlayer(Session session, String json) throws IOException {
-    	if(session.getUserProperties().get(Constants.USERNAME) != null) {
-    		return;		//already seen this user before on this socket
-    	}
+
     	JsonObject msg = Json.createReader(new StringReader(json)).readObject();
 		String username = Message.getValue(msg.get(Constants.USERNAME));
 		String userid = Message.getValue(msg.get(Constants.USERID));
 		
+    	System.out.println("*** Adding player "+userid+" from room "+room.getRoomId()+" via session "+session.getId());
 		room.addUserToRoom(userid, username);
-		room.command(userid, "look");
+		room.command(userid, "look");     
     }
     
-    private void removePlayer(Session session, String json) throws IOException {
+    private void removePlayer(Session session, String json) throws IOException {    	
     	JsonObject msg = Json.createReader(new StringReader(json)).readObject();
     	String userid = Message.getValue(msg.get(Constants.USERID));
+    	System.out.println("*** Removing player "+userid+" from room "+room.getRoomId()+" from session "+session.getId());
     	room.removeUserFromRoom(userid);
     }
